2015/9/25

+++ graph/MaximumFlow最大流.cpp
@@ -68,8 +68,8 @@ void init()
 }
 void add_edge(int from, int to, int cap)
 {
-    G[from].push_back((edge) {to, cap, G[to].size()});
-    G[to].push_back((edge) {from, 0, G[from].size() - 1 });
+    G[from].push_back((edge) {to, cap, (int) G[to].size()});
+    G[to].push_back((edge) {from, 0, (int) G[from].size() - 1 });
 }
 bool bfs(int s, int t)
 {
